Backend/PDF_Files/requirements.txt
apify_client==1.9.2
azure_storage==0.37.0
boto3==1.36.10
botocore==1.36.10
docling==2.24.0
fastapi==0.115.8
fitz==0.0.1.dev2
lxml==5.3.0
pdfplumber==0.11.5
Pillow==11.1.0
pymupdf==1.25.3
pytesseract==0.3.13
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
